Test values with different interest rates
Having determined values by calculating some values manually, we
now have tests to check that our calculator matches these calculations.

The tests check values before and after the rate change. They also
check the value where the rate is calculated for a period spanning
across the moment when the rate changes.

Rather than bifurcating based on dates, we map through a date range
and calculate the interest on each day. We also abstract the condition
on which rate to use into a single place. We only need to do this once,
and we also isoate it from the rest of the code. This also means we can
accommodate for further rate changes in the future more easily.